pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "canvas-lms-automation"
version = "2.0.0"
description = "Canvas LMS automation with PyQt6 GUI"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28.0",
    "PyQt6>=6.4.0",
    "selenium>=4.0.0",
    "pyotp>=2.8.0",
    "lxml>=4.9.0",
    "google-genai>=0.3.0",
    "anthropic>=0.18.0",
    "python-docx>=0.8.11",
    "Pillow>=9.0.0",
    "pypdf>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "func*", "gui*", "login*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
